requirements.txt
einops==0.6.0
grad-cam==1.4.8
hydra-colorlog==1.2.0
hydra-core==1.3.2
hydra-ax-sweeper==1.2.0
matplotlib==3.8.3
monai==1.3.0
opencv-python==4.7.0.72
openpyxl==3.1.2
pandas==2.0.1
Pillow==9.5.0
scikit-image==0.20.0
scikit-learn==1.2.2
scipy==1.9.1
seaborn==0.12.2
SimpleITK==2.2.1
tensorboard==2.12.0
timm==0.9.2
torchmetrics==1.2.0
transformers==4.36.1
